refactor: move prayRequestDto response package
